+++ CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.20.3] - 2025-10-21
+
+### New Feature
+
+- [#149](https://github.com/zpix1/yt-anti-translate/issues/149) Untranslate auto-play (autonav) video inside the player
+
+### Fixed
+
+- Fix [#148](https://github.com/zpix1/yt-anti-translate/issues/148) The mobile video thumbnail in a playlist are set to the first
+- Fix [#151](https://github.com/zpix1/yt-anti-translate/issues/151) searchbar error (polluted search history)
+- Fix [#120](https://github.com/zpix1/yt-anti-translate/issues/120) Chapters tooltip not untranslated in Brave Browser
+- Add some new youtube video selectors
+- Fix channel branding whitelist on search page
+
 ## [1.20.1] - 2025-10-14
 
 ### Fixed